Move new tests up into a `tests` subdirectory
The new tests of `gix_testtools::Env` are effectively end-to-end
tests, since they are involve modifying the actual environment, and
more importantly they are only testing the public surface of the
crate (`Env` is public), and therefore need not be inside any
non-test module.

So this moves from residing inside the nested `tests::env` module
within `tests/tools/src/lib.rs`, into the newly created
`tests/tools/tests/env.rs`.

(As expected, they still all pass, and when the fix in
`tests/tools/src/lib.rs` is temporarily undone, the `overlapping_*`
tests fail again, confirming they still work as regression tests.)
